[PagePartBundle][Tests] fix namespace tests and remove old mocks
