+++ src/Kunstmaan/PagePartBundle/Tests/_support/Helper/Unit.php
@@ -1,5 +1,5 @@
 <?php
-namespace PagePartBundle\Helper;
+namespace Kunstmaan\PagePartBundle\Tests\Helper;
 
 // here you can define custom actions
 // all public methods declared in helper class will be available in $I
