+++ src/Kunstmaan/PagePartBundle/Tests/_support/_generated/UnitTesterActions.php
@@ -1,5 +1,5 @@
-<?php  //[STAMP] 78ee096a68f63bd6a3edb6e7ad138714
-namespace PagePartBundle\_generated;
+<?php  //[STAMP] f09a3c18fea5eaf4fa70946f39696f53
+namespace Kunstmaan\PagePartBundle\Tests\_generated;
 
 // This class was automatically generated by build task
 // You should not change it manually as it will be overwritten on next build
