+++ src/Kunstmaan/PagePartBundle/Tests/unit.suite.yml
@@ -1,0 +1,5 @@
+actor: UnitTester
+modules:
+  enabled:
+  - Asserts
+  - Kunstmaan\PagePartBundle\Tests\Helper\Unit
