+++ src/Kunstmaan/PagePartBundle/Tests/unit/Form/HeaderPagePartAdminTypeTest.php
@@ -2,15 +2,16 @@
 
 namespace Kunstmaan\PagePartBundle\Tests\Form;
 
-use Kunstmaan\PagePartBundle\Form\LinePagePartAdminType;
+use Kunstmaan\PagePartBundle\Form\HeaderPagePartAdminType;
+use Kunstmaan\PagePartBundle\Tests\unit\Form\PagePartAdminTypeTestCase;
 
 /**
- * Generated by PHPUnit_SkeletonGenerator on 2012-08-20 at 13:19:20.
+ * Generated by PHPUnit_SkeletonGenerator on 2012-08-20 at 13:19:22.
  */
-class LinePagePartAdminTypeTest extends PagePartAdminTypeTestCase
+class HeaderPagePartAdminTypeTest extends PagePartAdminTypeTestCase
 {
     /**
-     * @var LinePagePartAdminType
+     * @var HeaderPagePartAdminType
      */
     protected $object;
 
@@ -21,18 +22,20 @@ class LinePagePartAdminTypeTest extends PagePartAdminTypeTestCase
     protected function setUp()
     {
         parent::setUp();
-        $this->object = new LinePagePartAdminType();
+        $this->object = new HeaderPagePartAdminType();
     }
 
     public function testBuildForm()
     {
         $this->object->buildForm($this->builder, array());
+        $this->builder->get('niv');
+        $this->builder->get('title');
     }
 
     public function testConfigureOptions()
     {
         $this->object->configureOptions($this->resolver);
         $resolve = $this->resolver->resolve();
-        $this->assertEquals($resolve["data_class"], 'Kunstmaan\PagePartBundle\Entity\LinePagePart');
+        $this->assertEquals($resolve["data_class"], 'Kunstmaan\PagePartBundle\Entity\HeaderPagePart');
     }
 }
