+++ src/Kunstmaan/PagePartBundle/Tests/unit/Form/LinePagePartAdminTypeTest.php
@@ -2,16 +2,16 @@
 
 namespace Kunstmaan\PagePartBundle\Tests\Form;
 
-use Kunstmaan\PagePartBundle\Form\TocPagePartAdminType;
+use Kunstmaan\PagePartBundle\Form\LinePagePartAdminType;
+use Kunstmaan\PagePartBundle\Tests\unit\Form\PagePartAdminTypeTestCase;
 
 /**
- * Generated by PHPUnit_SkeletonGenerator on 2012-08-20 at 12:29:42.
+ * Generated by PHPUnit_SkeletonGenerator on 2012-08-20 at 13:19:20.
  */
-class TocPagePartAdminTypeTest extends PagePartAdminTypeTestCase
+class LinePagePartAdminTypeTest extends PagePartAdminTypeTestCase
 {
-
     /**
-     * @var TocPagePartAdminType
+     * @var LinePagePartAdminType
      */
     protected $object;
 
@@ -22,7 +22,7 @@ class TocPagePartAdminTypeTest extends PagePartAdminTypeTestCase
     protected function setUp()
     {
         parent::setUp();
-        $this->object = new TocPagePartAdminType();
+        $this->object = new LinePagePartAdminType();
     }
 
     public function testBuildForm()
@@ -34,6 +34,6 @@ class TocPagePartAdminTypeTest extends PagePartAdminTypeTestCase
     {
         $this->object->configureOptions($this->resolver);
         $resolve = $this->resolver->resolve();
-        $this->assertEquals($resolve["data_class"], 'Kunstmaan\PagePartBundle\Entity\TocPagePart');
+        $this->assertEquals($resolve["data_class"], 'Kunstmaan\PagePartBundle\Entity\LinePagePart');
     }
 }
