+++ src/Kunstmaan/PagePartBundle/Tests/unit/Form/LinkPagePartAdminTypeTest.php
@@ -2,15 +2,16 @@
 
 namespace Kunstmaan\PagePartBundle\Tests\Form;
 
-use Kunstmaan\PagePartBundle\Form\HeaderPagePartAdminType;
+use Kunstmaan\PagePartBundle\Form\LinkPagePartAdminType;
+use Kunstmaan\PagePartBundle\Tests\unit\Form\PagePartAdminTypeTestCase;
 
 /**
- * Generated by PHPUnit_SkeletonGenerator on 2012-08-20 at 13:19:22.
+ * Generated by PHPUnit_SkeletonGenerator on 2012-08-20 at 13:19:18.
  */
-class HeaderPagePartAdminTypeTest extends PagePartAdminTypeTestCase
+class LinkPagePartAdminTypeTest extends PagePartAdminTypeTestCase
 {
     /**
-     * @var HeaderPagePartAdminType
+     * @var LinkPagePartAdminType
      */
     protected $object;
 
@@ -21,20 +22,19 @@ class HeaderPagePartAdminTypeTest extends PagePartAdminTypeTestCase
     protected function setUp()
     {
         parent::setUp();
-        $this->object = new HeaderPagePartAdminType();
+        $this->object = new LinkPagePartAdminType();
     }
 
     public function testBuildForm()
     {
         $this->object->buildForm($this->builder, array());
-        $this->builder->get('niv');
-        $this->builder->get('title');
+        $this->builder->get('url');
     }
 
     public function testConfigureOptions()
     {
         $this->object->configureOptions($this->resolver);
         $resolve = $this->resolver->resolve();
-        $this->assertEquals($resolve["data_class"], 'Kunstmaan\PagePartBundle\Entity\HeaderPagePart');
+        $this->assertEquals($resolve["data_class"], 'Kunstmaan\PagePartBundle\Entity\LinkPagePart');
     }
 }
