+++ src/Kunstmaan/PagePartBundle/Tests/unit/Form/ToTopPagePartAdminTypeTest.php
@@ -3,6 +3,7 @@
 namespace Kunstmaan\PagePartBundle\Tests\Form;
 
 use Kunstmaan\PagePartBundle\Form\ToTopPagePartAdminType;
+use Kunstmaan\PagePartBundle\Tests\unit\Form\PagePartAdminTypeTestCase;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator on 2012-08-20 at 12:03:51.
