+++ src/Kunstmaan/PagePartBundle/Tests/unit/Form/TocPagePartAdminTypeTest.php
@@ -2,15 +2,17 @@
 
 namespace Kunstmaan\PagePartBundle\Tests\Form;
 
-use Kunstmaan\PagePartBundle\Form\LinkPagePartAdminType;
+use Kunstmaan\PagePartBundle\Form\TocPagePartAdminType;
+use Kunstmaan\PagePartBundle\Tests\unit\Form\PagePartAdminTypeTestCase;
 
 /**
- * Generated by PHPUnit_SkeletonGenerator on 2012-08-20 at 13:19:18.
+ * Generated by PHPUnit_SkeletonGenerator on 2012-08-20 at 12:29:42.
  */
-class LinkPagePartAdminTypeTest extends PagePartAdminTypeTestCase
+class TocPagePartAdminTypeTest extends PagePartAdminTypeTestCase
 {
+
     /**
-     * @var LinkPagePartAdminType
+     * @var TocPagePartAdminType
      */
     protected $object;
 
@@ -21,19 +23,18 @@ class LinkPagePartAdminTypeTest extends PagePartAdminTypeTestCase
     protected function setUp()
     {
         parent::setUp();
-        $this->object = new LinkPagePartAdminType();
+        $this->object = new TocPagePartAdminType();
     }
 
     public function testBuildForm()
     {
         $this->object->buildForm($this->builder, array());
-        $this->builder->get('url');
     }
 
     public function testConfigureOptions()
     {
         $this->object->configureOptions($this->resolver);
         $resolve = $this->resolver->resolve();
-        $this->assertEquals($resolve["data_class"], 'Kunstmaan\PagePartBundle\Entity\LinkPagePart');
+        $this->assertEquals($resolve["data_class"], 'Kunstmaan\PagePartBundle\Entity\TocPagePart');
     }
 }
