+++ src/Kunstmaan/PagePartBundle/Tests/unit/PagePartConfigurationReader/Resources/config/pageparts/main.yml
